refactor: Split up `Primer.Core` module
The original intention here was to be able to put `Primitives.defType` somewhere more sensible. But as a bonus, this _should_ result in better compilation times, particularly incrementally.

We also take the opportunity to simplify the definition of `primConName`, now that it is in a module where it can access `tChar` etc.
