+++ primer/test/Tests/Pretty.hs
@@ -8,8 +8,9 @@ import Data.Text qualified as T
 import Prettyprinter (defaultLayoutOptions, layoutSmart, line)
 import Prettyprinter.Internal.Type (Doc)
 import Prettyprinter.Render.Terminal (AnsiStyle, renderStrict)
-import Primer.Core (ASTDef (..), Def (..), GVarName, GlobalName (baseName), ModuleName, mkSimpleModuleName)
+import Primer.Core (GVarName, GlobalName (baseName), ModuleName, mkSimpleModuleName)
 import Primer.Core.DSL (S, create')
+import Primer.Def (ASTDef (..), Def (..))
 import Primer.Examples (comprehensive, not)
 import Primer.Name (unName)
 import Primer.Pretty (PrettyOptions (..), compact, prettyExpr, prettyType, sparse)
